requirements.txt
appnope==0.1.0
attrs==18.1.0
backcall==0.1.0
decorator==4.3.0
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
more-itertools==4.1.0
parso==0.2.0
peewee==3.3.4
pexpect==4.5.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
py==1.5.3
Pygments==2.2.0
pytest==3.5.1
pytest-mock==1.10.0
python-chess==0.23.5
python-decouple==3.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
